Add delay to fix possible I2C write fail
For the following commands, an I2C write fail might occur when the
sensor is still processing the command and a new command is sent:

 - sps30_start_measurement
 - sps30_stop_measurement
 - sps30_set_fan_auto_cleaning_interval
 - sps30_set_fan_auto_cleaning_interval_days
 - sps30_start_manual_fan_cleaning
 - sps30_reset

A delay is added to these commands (except sps30_reset) to give the
sensor enough time to process these commands. For sps30_reset, the
caller should not interact with the sensor for the specified period (see
header file).